mapper

+++ lazy-vm/src/main/java/com/lazy/vm/domain/Exam.java
@@ -44,7 +44,7 @@ public class Exam extends BaseEntity
 
     /** 考试总分 */
     @Excel(name = "考试总分")
-    private Long totalScore;
+    private Integer totalScore;
 
     /** 限考次数 */
     @Excel(name = "限考次数")
@@ -223,11 +223,11 @@ public class Exam extends BaseEntity
         this.updateTime = updateTime;
     }
 
-    public Long getTotalScore() {
+    public Integer getTotalScore() {
         return totalScore;
     }
 
-    public void setTotalScore(Long totalScore) {
+    public void setTotalScore(Integer totalScore) {
         this.totalScore = totalScore;
     }
